Fix cleanup code
Traceback (most recent call last):
  File "/usr/sbin/univention-management-console-module", line 121, in <module>
    notifier.loop()
  File "/usr/lib/pymodules/python2.7/notifier/nf_generic.py", line 287, in loop
    step()
  File "/usr/lib/pymodules/python2.7/notifier/nf_generic.py", line 243, in step
    if not timer[ CALLBACK ]():
  File "/usr/lib/pymodules/python2.7/univention/management/console/protocol/modserver.py", line 140, in _timer
    self._timed_out()
  File "/usr/lib/pymodules/python2.7/univention/management/console/protocol/modserver.py", line 150, in _timed_out
    self.exit()
  File "/usr/lib/pymodules/python2.7/univention/management/console/protocol/server.py", line 392, in exit
    os.unlink(self.__unix)
OSError: [Errno 2] Datei oder Verzeichnis nicht gefunden: '/var/run/univention-management-console/1621-1570639160845.socket'

+++ management/univention-management-console/src/univention/management/console/protocol/server.py
@@ -389,7 +389,8 @@ class Server(signals.Provider):
 			self.__realsocket.close()
 			self.__realsocket = None
 		if self.__unix:
-			os.unlink(self.__unix)
+			if os.path.exists(self.__unix):
+				os.unlink(self.__unix)
 			self.__unix = None
 
 		self.__bucket = None
